fix(HardwareSerial): fix pin remapping in begin()
The pin remapping functions have to be called as early as possible in
the begin() function, to immediately convert the input parameters to the
GPIO numbers used everywhere in the core.

This issue has always been dormant since the introduction of pin
remapping in 9b4622d, but was exposed by the recent UART pin detach
support, which actually disabled the original Serial0 pins.

Move the pin remapping function calls earlier in the begin() function to
fix this issue.