platform.txt: use iwithprefixbefore to shorten the command lines
The command line for the ESP32 compilers is very long, and it is
causing problems on Windows systems, which enforce  a hard 32kbytes
line limit.

This patch uses the iprefix/iwithprefixbefore options to shorten the
compiler command line by changing

	"-I{compiler.sdk.path}/include/aaa"
	"-I{compiler.sdk.path}/include/bbb"
	"-I{compiler.sdk.path}/include/ccc" ...

to:

	-iprefix "{compiler.sdk.path}/include/"
	-iwithprefixbefore aaa
	-iwithprefixbefore bbb
	-iwithprefixbefore ccc ...

Some targets have up to 200 paths so this can save A LOT of space.
On a sample run this patch shortened a single compile command from
28777 bytes to 10728.